+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This Changelog only applies to notable changes to `helios.js` on the main branch
 * Script purpose keywords changed to 'testing', 'spending' and 'minting'
 * Macro-like builtins whichs are only allowed after 'main': ValidatorHash::CURRENT, MintingPolicyHash::CURRENT, ScriptContext::new, Tx::new, TxID::CURRENT, TxInput::new, TxOutput::new
 * Generally usable constructors: TxId::new, Address::new, Credential::new_pubkey, Credential::new_validator, StakingCredential::new_hash, StakingCredential::new_ptr
+* First argument of AssetClass::new must be MintingPolicyHash (was ByteArray previously)
 
 # 2022/08/17
 * Cost calculation of a plutus-core program run